Develop

+++ src/components/MyAccount/DeleteAccount.js
@@ -1,6 +1,6 @@
-/* eslint-disable no-console */
 import './MyAccount.scss';
 import { useSelector, useDispatch } from 'react-redux';
+import { toast } from 'react-toastify';
 import { deleteAccount } from '../../actions/user';
 import api from '../../api/api';
 
@@ -18,43 +18,63 @@ function DeleteAccount() {
     api
       .delete(`/users/${userId}`)
       .then((response) => {
-        console.log(response);
+        // console.log(response);
         localStorage.setItem('token', '');
         localStorage.setItem('nickname', '');
         localStorage.setItem('id', '');
-        dispatch(deleteAccount());
+        const confirmation = window.confirm('Confirmez-vous la suppression du compte ?');
+        if (confirmation) {
+          localStorage.setItem('token', '');
+          localStorage.setItem('nickname', '');
+          localStorage.setItem('id', '');
+          dispatch(deleteAccount());
+          toast.success('Suppression réussie', {
+            position: 'top-right',
+            autoClose: 3000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: 'colored',
+          });
+          // Différer la redirection de 3.5 secondes (3500 millisecondes)
+          setTimeout(() => {
+            window.location.href = '/';
+          }, 3500);
+        }
       })
       .catch((error) => {
         console.log(error);
+        toast.error('Echec de la suppression', {
+          position: 'top-right',
+          autoClose: 3000,
+          hideProgressBar: false,
+          closeOnClick: true,
+          pauseOnHover: true,
+          draggable: true,
+          progress: undefined,
+          theme: 'colored',
+        });
       });
   };
 
   return (
-    <div className="MyAccount-Show">
-      <div className="MyAccount-Delete">
-        <h1>Supprimer mon compte</h1>
-        <p className="MyAccount-red">Attention, en cliquant sur le bouton "supprimer mon compte" ci dessous,
-          cela entrainement la suppression définitive de votre compte et vous perdrez
-          toute votre progression dans le jeu.
-        </p>
-        <p className="MyAccount-red">
-          Si vous souhaitez rejouer à nouveau, il vous faudra créer un nouveau compte.
-        </p>
+    <div className="MyAccount-Delete">
+      <p className="MyAccount-Delete-warning">Attention, en cliquant sur le bouton ci-dessous,
+        cela entraînera la suppression définitive de votre compte.
+      </p>
+      <p className="MyAccount-Delete-warning">Vous devrez en recréer un pour pouvoir jouer.
+      </p>
 
-        <h1 className="MyAccount-question">Souhaitez-vous supprimer votre compte ?</h1>
-        <p className="MyAccount-troll">
-          Souhaitez-vous vraiment nous briser le coeur ?
-          Parce que oui ! C'est ce que vous risquez de faire en osant cliquer sur ce bouton,
-          le roi des Cookie pourrait venir vous dévorer
-          pour se venger ! Vous n'avez pas peur ? Peur du Cookie vengeur ?
-        </p>
-        <button
-          type="button"
-          onClick={() => deleteUserAccount()}
-        >
-          - Non j'ai pas peur moi et oui je souhaite supprimer mon compte
-        </button>
-      </div>
+      <h2 className="MyAccount-Delete-question">Souhaitez-vous supprimer votre compte ?</h2>
+      <button
+        className="MyAccount-Delete-button"
+        type="button"
+        onClick={() => deleteUserAccount()}
+      >
+        J'ai bien compris et je souhaite supprimer mon compte.
+      </button>
     </div>
   );
 }
